requirements.txt
fastapi
uvicorn
sqlalchemy
google-auth
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
groq
scholarly
python-docx
python-telegram-bot